first split - moved NeuronUpdateGroupMerged out into its own module
