+++ src/genn/genn/code_generator/backendBase.cc
@@ -6,6 +6,7 @@
 
 // GeNN code generator includes
 #include "code_generator/groupMerged.h"
+#include "code_generator/neuronUpdateGroupMerged.h"
 
 // Macro for simplifying defining type sizes
 #define TYPE(T) {#T, {sizeof(T), std::to_string(std::numeric_limits<T>::lowest())}}
